{{cookiecutter.repo_name}}/requirements/development.txt
# Local development dependencies go here
-r common.txt
selenium
coverage
Sphinx

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar
